proc: implement hardware backed Memory Guard API
Memory-Guard is a ninex specific API designed for sanitizing and
verifying usermode pointers, while also limiting access to usermode
data. It does this by running a series of checks on pointers passed to
the kernel, while using hardware extensions, such as SMAP to achive this
effect. Some other changes were made in this commit as well, such as...

  * Recognizing the presence of the X-DSDT field in the FADT and
    utilizing it
  * Adding more syscalls, such as ioctl() and getcwd()
  * Reclaming memory from exited processes, and fixing fatal memory bug
    within scheduler where a idle CPU's address space would get
    reclaimed while CPU idles